requirements.txt
requests
pytest
flask
python-dotenv